refactor: split main functionalities
